==> notebooks/01_transformaciones.ipynb <==
# 01 - Proyecto Individual: Transformaciones

## Carga de las librerías necesarias.

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

## Carga e inspección del dataset `games`.

from google.colab import drive
drive.mount('/content/drive')

%cd /content/drive/MyDrive/AI Projects/PI01

!ls datasets/

!head -n 4 datasets/output_steam_games.json

games_df = pd.read_json('datasets/output_steam_games.json', lines=True)
games_df.info()

np.random.seed(42)
games_df.sample(n=10)

Eliminamos los registros con el campo `id` nulo.

games_df.shape

games_df.dropna(subset=['id'], inplace=True)

## Revisamos cada campo del dataset.

games_df.shape

games_df.isnull().sum()

games_df.notnull().sum()

### Campo `publisher`.

np.random.seed(42)
games_df.publisher.sample(n=10)

### Campo `genres`.

np.random.seed(42)
games_df.genres.sample(n=10)

### Campo `app_name`.

np.random.seed(42)
games_df.app_name.sample(n=10)

### Campo `title`.

np.random.seed(42)
games_df.title.sample(n=10)

### Campo `url`.

np.random.seed(42)
games_df.url.sample(n=10)

### Campo `release_date`.

np.random.seed(42)
games_df.release_date.sample(n=10)

games_df.loc[:, 'release_date'] = games_df.release_date.map(lambda x: pd.to_datetime(x, errors='coerce'))

np.random.seed(42)
games_df.release_date.sample(n=10)

### Campo `tags`.

np.random.seed(42)
games_df.tags.sample(n=10)

### Campo `reviews_url`.

np.random.seed(42)
games_df.reviews_url.sample(n=10)

### Campo `specs`.

np.random.seed(42)
games_df.specs.sample(n=10)

### Campo `price`.

np.random.seed(42)
games_df.price.sample(n=10)

games_df.price = games_df.price.map(lambda x: pd.to_numeric(x, errors='coerce'))

np.random.seed(42)
games_df.price.sample(n=10)

### Campo `early_access`.

np.random.seed(42)
games_df.early_access.sample(n=10)

np.random.seed(42)
games_df.early_access.value_counts()

### Campo `id`.

np.random.seed(42)
games_df.id.sample(n=10)

games_df.id = games_df.id.astype(int)

np.random.seed(42)
games_df.id.sample(n=10)

### Campo `developer`.

np.random.seed(42)
games_df.developer.sample(n=10)

## Carga e inspección del dataset `reviews`.

!head -n 10 datasets/australian_user_reviews.json

from ast import literal_eval

with open("datasets/australian_user_reviews.json", 'r') as f:
  data = [literal_eval(line.strip()) for line in f]

data[:2]

reviews_df = pd.DataFrame(data)

reviews_df.info()

reviews_df.shape

reviews_df.notnull().sum()

reviews_df.isnull().sum()

### Campo `user_id`.

np.random.seed(42)
reviews_df.user_id.sample(n=20)

### Campo `user_url`.

np.random.seed(42)
reviews_df.user_url.sample(n=10)

### Campo `user_review`.

np.random.seed(42)
reviews_df.reviews.sample(n=10)

reviews_df.reviews[9841]

reviews_df.reviews[2222]

### Creamos los campos `item_id`, `recommend` y `review`.

def helper_fn(list_):
  recommends = []
  for item in list_:
    recommends.append(item['recommend'])
  return recommends

reviews_df['recommend'] = reviews_df.reviews.map(helper_fn)

np.random.seed(42)
reviews_df.recommend.sample(n=10)

def helper_fn2(list_):
  reviews = []
  for item in list_:
    reviews.append(item['review'])
  return reviews

reviews_df['review'] = reviews_df.reviews.map(helper_fn2)

np.random.seed(42)
reviews_df.review.sample(n=10)

def helper_fn3(list_):
  item_ids = []
  for item in list_:
    item_ids.append(item['item_id'])
  return item_ids

reviews_df['item_id'] = reviews_df.reviews.map(helper_fn3)

np.random.seed(42)
reviews_df.item_id.sample(n=10)

## Cargamos el dataset `user_items`.

from ast import literal_eval

with open("datasets/australian_users_items.json", 'r') as f:
  data = [literal_eval(line.strip()) for line in f]

items_df = pd.DataFrame(data)

items_df.info()

del data

import gc
gc.collect()

items_df.shape

items_df.isnull().sum()

items_df.notnull().sum()

### Campo `user_id`.

np.random.seed(42)
items_df.user_id.sample(n=10)

### Campo `items_count`.

np.random.seed(42)
items_df.items_count.sample(n=10)

### Campo `steam_id`.

np.random.seed(42)
items_df.steam_id.sample(n=10)

items_df.steam_id = pd.to_numeric(items_df.steam_id)

np.random.seed(42)
items_df.steam_id.sample(n=10)

### Campo `user_url`.

np.random.seed(42)
items_df.user_url.sample(n=10)

### Campo `items`.

np.random.seed(42)
items_df["items"].sample(n=10)

items_df['items'][64750]

### Creamos los campos `item_id`, `item_name`, `playtime_forever`.

def helper_fn(list_):
  item_ids = []
  for item in list_:
    item_ids.append(item['item_id'])
  return item_ids

items_df['item_id'] = items_df['items'].map(helper_fn)
items_df['item_id'].head()

def helper_fn(list_):
  item_playtimes = []
  for item in list_:
    item_playtimes.append(item['playtime_forever'])
  return item_playtimes

items_df['item_playtime'] = items_df['items'].map(helper_fn)
items_df['item_playtime'].head()

def helper_fn(list_):
  item_names = []
  for item in list_:
    item_names.append(item['item_name'])
  return item_names

items_df['item_name'] = items_df['items'].map(helper_fn)
items_df['item_name'].head()

## Guardamos los datasets

games_columns = ['id', 'genres', 'release_date', 'tags', 'specs', 'developer']
games_df[games_columns].to_parquet('datasets/games.parquet', index=False)

reviews_columns = ['user_id', 'item_id', 'recommend', 'review']
reviews_df = reviews_df[reviews_columns].explode(['recommend', 'review', 'item_id'])

reviews_df.info()

reviews_df.to_parquet('datasets/reviews.parquet', index=False)

items_df.info()

items_columns = ['user_id', 'steam_id', 'item_id', 'item_name', 'item_playtime']
items_df[items_columns].explode(['item_id', 'item_name', 'item_playtime']).to_parquet('datasets/items.parquet', index=False)
