==> notebooks/02_feature_engineering.ipynb <==
# 02 - Proyecto individual: Feature Engineering

## Instalamos las librrerías requeridas.

Para realizar el análisis de sentimientos del los reviews del dataset `reviews`.

!pip install transformers

## Montamos la partición de Google Drive.

Para acceder al dataset preprocesado `reviews`.

%cd /content/drive/MyDrive/AI Projects/PI01

## Cargamos las liberías necesarias.

import pandas as pd
import numpy as np
import transformers

## Cargamos el dataset `reviews`.

reviews_df = pd.read_parquet('datasets/reviews.parquet')

reviews_df.info()

reviews_df.head()

### Verificamos que no hallan valores nulos.

reviews_df.review.isnull().sum()

reviews_df = reviews_df.query('review.notnull()')

reviews_df.review.isnull().sum()

## Instanciamos el clasificador de setimientos.

sentiment_pipeline = transformers.pipeline(model='mrm8488/distilroberta-finetuned-financial-news-sentiment-analysis',
                                           max_length=512, truncation=True, device=0)

### Testeamos el modelo clasificador antes

np.random.seed(42)
reviews_sample = reviews_df.review.sample(5).tolist()
reviews_sample

prediction = sentiment_pipeline(reviews_sample)
prediction

### Aplicamos el analizador de sentimientos a todo el campo `review`.

reviews_df['sentiment'] = sentiment_pipeline(reviews_df.review.tolist())

def helper_fn(item):
  polarity = 0
  label = item['label']
  if label == 'positive':
    polarity = 2
  elif label == 'negative':
    polarity = 0
  elif label == 'neutral':
    polarity = 1

  return polarity

reviews_df['sentiment'] = reviews_df.sentiment.map(helper_fn)
reviews_df['sentiment'].value_counts()

### Guardamos el dataset actualizado.

reviews_df.info()

reviews_columns = ['item_id', 'user_id', 'recommend', 'sentiment']

reviews_df[reviews_columns].to_parquet('datasets/reviews_stt.parquet', index=False)
