==> notebooks/03_api_dev_and_test.ipynb <==
# Proyecto Individual - Desarrollo y prueba de las funciones del API.

import pandas as pd
import numpy as np

from google.colab import drive
drive.mount('/content/drive')

%cd "/content/drive/MyDrive/AI Projects/PI01"

## Cargamos los datasets.

games_df = pd.read_parquet('datasets/games.parquet')
reviews_df = pd.read_parquet('datasets/reviews_stt.parquet')
items_df = pd.read_parquet('datasets/items.parquet')

### Modificamos el dataset `games`.

games_df.info()

games_df.sample(n=5)

games_df.drop_duplicates(subset=['id'], inplace=True)
games_df.shape

#### Explotamos el campo `genres` del dataset `games`.

games_df = games_df.explode(column=['genres'], ignore_index=True)
games_df.head(n=5)

games_df.columns = ['id', 'genre', 'release_date', 'tags', 'specs', 'developer']

games_df.info()

### Modificamos el dataset `reviews`.

reviews_df.info()

reviews_df.item_id = pd.to_numeric(reviews_df.item_id, errors='coerce')

reviews_df.info()

### Modificamos el dataset `items`.

items_df.info()

items_df.drop('steam_id', axis=1, inplace=True)

items_df.dropna(subset=['item_id'], inplace=True)

items_df.item_id.shape

items_df.item_id = pd.to_numeric(items_df.item_id, errors='coerce')

items_df.info()

items_df.sample(n=5)

items_df.columns

items_df.columns = ['user_id', 'id', 'item_name', 'item_playtime']

## Obtener el año con más horas jugadas por género.

Creamos un dataset que junte los daros de `games` e `items` por el campo `id`.

games_df.shape

items_df.shape

items_games_df = items_df.join(games_df[['id', 'genre', 'release_date']].drop_duplicates(subset=['id']).set_index('id'),
                               on=['id'], how='left')
items_games_df.release_date = items_games_df.release_date.dt.year
items_games_df.item_playtime = (items_games_df.item_playtime/60)
items_games_df.sample(10)

items_games_df.shape

def playTimeByGenre(genre:str):
    year = items_games_df.query('genre == @genre').groupby('release_date').agg({'item_playtime': 'sum'}).sort_values(by='item_playtime', ascending=False).index[0]

    return {f"Año con más horas jugas para el género {genre}": int(year)}

Probamos la función.

genres = ['Action', 'Indie', 'RPG', 'Simulation']

for genre in genres:
    print(playTimeByGenre(genre))

## Obtener el usuario con más horas jugadas acumuladas por año.

def userForGenre(genre:str):
    user_id = items_games_df.query('genre == @genre').groupby('user_id').agg({'item_playtime': 'sum'}).sort_values(by='item_playtime', ascending=False).index[0]
    temp = items_games_df.query('user_id == @user_id & genre == @genre').groupby('release_date').agg({'item_playtime': 'sum'}).to_dict()['item_playtime']
    playtime_by_year = [{"Año": int(year), "Horas": round(time, 2)} for year, time in temp.items() if time > 0]

    return {f"Usuario con más horas jugadas en el género {genre}": user_id,
            "Horas jugadas": playtime_by_year}

Probamos la función.

genres = ['Action', 'Indie', 'RPG', 'Simulation']

for genre in genres:
    print(userForGenre(genre))

## Top 3 de los juegos más recomendados en un año dado.

games_df.columns

reviews_df.columns

items_df.columns

reviews_df.columns = ['id', 'user_id', 'recommend', 'sentiment']

games_df.shape, reviews_df.shape

reviews_games_items_df = reviews_df.join(games_df[['id', 'release_date']].drop_duplicates().set_index('id'),
                on=['id'], how='left')
reviews_games_items_df.release_date = reviews_games_items_df.release_date.dt.year
reviews_games_items_df = reviews_games_items_df.join(items_df[['id', 'item_name']].drop_duplicates().set_index('id'),
                                                     on=['id'], how='left')
reviews_games_items_df.sample(n=10)

def userRecommended(anio: int):
  top_3 = reviews_games_items_df.query("release_date==@anio & recommend==True & sentiment>=1").groupby('item_name').agg({'recommend': 'count'}).sort_values(by='recommend', ascending=False).iloc[:3].to_dict()['recommend']
  top_3 = list(top_3.keys())
  return {f'Puesto {idx}': title for idx, title in enumerate(top_3, start=1)}

for anio in range(2005, 2015):
    print(userRecommended(anio))

## Top 3 de los juegos menos recommendados en un año dado.

def userNotRecommended(anio: int):
  top_3 = reviews_games_items_df.query("release_date==@anio & recommend==False & sentiment<1").groupby('item_name').agg({'recommend': 'count'}).sort_values(by='recommend', ascending=False).iloc[:3].to_dict()['recommend']
  top_3 = list(top_3.keys())
  return {f'Puesto {idx}': title for idx, title in enumerate(top_3, start=1)}

for anio in range(2005, 2015):
    print(userNotRecommended(anio))

## Obtener la cantidad de reseñas por categoria de acuerdo con un año dado.

def sentimentPolarityByYear(anio:int):
  polarities = reviews_games_items_df.query('release_date==@anio').sentiment.value_counts().to_dict()
  return {'Negative': polarities[0], 'Neutral': polarities[1], 'Positive': polarities[2]}

for anio in range(2005, 2015):
    print(sentimentPolarityByYear(anio))

## Guardamos los datasets creados.

A usarse al hacer consultas por la API.

items_games_df.to_parquet('datasets/items_games.parquet', index=False)
reviews_games_items_df.to_parquet('datasets/reviews_games_items.parquet', index=False)
