==> notebooks/05_recommender_system.ipynb <==
# 05 - Proyecto Individual: Sistema de recomendación.

## Carga de las librerías requeridas.

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

## Carga del dataset `ml_games`.

%cd /content/drive/MyDrive/AI Projects/PI01

ml_games = pd.read_parquet('datasets/ml_games.parquet').reset_index(drop=True)
ml_games.head(n=10)

ml_games.shape

Reduciremos el dataset para poder almacenar la matriz de similitud que se creará más adelante.

np.random.seed(42)
ml_games = ml_games.sample(frac=.90)

ml_games.shape

## Vectorización del campo `bag_of_words`.

from sklearn.feature_extraction.text import TfidfVectorizer

vectorizer = TfidfVectorizer()
vectorizer.fit(ml_games.bag_of_words)

vec_words = vectorizer.transform(ml_games.bag_of_words)

vec_words.shape

### Cálculo de la matriz de similaridad.

Se usará el *coseno de similaridad*.

from sklearn.metrics.pairwise import linear_kernel

similarity_matrix = linear_kernel(vec_words, vec_words)

similarity_matrix.shape

similarity_matrix = similarity_matrix.astype('float16')

### Desarrollo de la función de recomendación.

def gameRecommendation(id):
  similarity_matrix_idx = ml_games.query('id == @id').index[0]
  similarity_matrix_row = similarity_matrix[similarity_matrix_idx]
  game_row_idxs = similarity_matrix_row.argsort()[-6:-1][::-1]

  ref_game = ml_games.query('id == @id')[['id', 'item_name']].to_dict(orient='records')[0]
  top_5 = ml_games[['id', 'item_name']].iloc[game_row_idxs].to_dict(orient='records')
  return {'game':ref_game, 'top_5': top_5}

gameRecommendation(130)

## Guardado de la matriz de similitud y el dataset `ml_games`.

np.savez_compressed('datasets/similarity_matrix.npz', similarity_matrix)

!ls -sh datasets/similarity_matrix.npz

ml_games.to_parquet('datasets/ml_games.parquet', index=True)
